Linear Time (?) BWTd.c, updated BWTe.c
Made a linear time BWT decoder using bucket sort.
Not sure if this is true (that the program’s performance is in O(n)
time), but its performance is much faster than the forward
transformation.

Going to have to comment and edit this properly so that it’s readable
and maintainable sometime in the future.

The BWT encoder has been updated so that it should work with bigger
files now, also more use of unsigned char for consistency.